+++ src/components/auth/signIn/SignInForm.scss
@@ -1,4 +1,4 @@
-@use '../styles/variables.scss';
+@use '@/styles/variables.scss';
 
 a {
   color: variables.$text-secondary;
