+++ src/layout/Auth.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import signInBg from "@/assets/signIn-bg.jpg"
-import './AuthLayout.scss';
+import '@/layout/Auth.scss';
 
 const AuthLayout = () => {
   return (
