+++ src/pages/auth/SignUp.tsx
@@ -1,4 +1,4 @@
-import SignUpForm from "@/components/SignUpForm";
+import SignUpForm from "@/components/auth/signup/SignUpForm";
 
 const SignUp: React.FC = () => {
